# docs/swagger.yaml
definitions:
  dto.CreateUpdateSubscribeDto:
    properties:
      end_date:
        type: string
      name:
        type: string
      price:
        type: integer
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.ReadSubscribeDto:
    properties:
      end_date:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: integer
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.SubscribeCostDto:
    properties:
      total:
        type: number
    type: object
info:
  contact: {}
paths:
  /subscribe:
    get:
      description: Get list of subscribes with pagination, filtering and sorting
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Sort fields, e.g. name:asc,price:desc
        in: query
        name: sort
        type: string
      - description: Filters in the form [{
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ReadSubscribeDto'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get list of subscribes
      tags:
      - Subscribe
    post:
      consumes:
      - application/json
      description: Create new subscribe
      parameters:
      - description: Subscribe data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUpdateSubscribeDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create new subscribe
      tags:
      - Subscribe
  /subscribe/{id}:
    delete:
      description: Delete subscribe by ID
      parameters:
      - description: Subscribe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete subscribe by ID
      tags:
      - Subscribe
    get:
      description: Get subscribe by ID
      parameters:
      - description: Subscribe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadSubscribeDto'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Get subscribe by ID
      tags:
      - Subscribe
    patch:
      consumes:
      - application/json
      description: Update subscribe by ID
      parameters:
      - description: Subscribe ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscribe data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUpdateSubscribeDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update subscribe by ID
      tags:
      - Subscribe
  /subscribes/cost:
    get:
      consumes:
      - application/json
      description: Returns the total subscription cost for a user with optional filters
        by subscription name and date range
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Subscription name
        in: query
        name: name
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscribeCostDto'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get total subscription cost
      tags:
      - subscriber-cost
swagger: "2.0"